code cleanup

+++ main_test.cpp
@@ -19,7 +19,7 @@ class thread_test {
     thread_test(log_container *log): log_(log) {}
 
     void writer_thread() {
-        for (unsigned int i = 0; i < 100; i++) {
+        for (uint64_t i = 0; i < 100; i++) {
             std::stringstream ss;
             ss << i;
             log_->append(string_record("hello_world: " + ss.str()));
